requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Flask-Migrate==4.0.5
requests==2.31.0
asyncio==3.4.3
Authlib==1.2.1
Flask-Session==0.5.0
pytest